Move position area declarations from fallbacks file
